feat: view controller -> table view cell 변경 (#3)
